add debug flag Fixes #3
